Extract ToolSelection and CustomToolFactory into tool_surface/selection.rs
